Fix sql

+++ alter2.sql
@@ -1,8 +1,8 @@
 CREATE TABLE `favorite` (
   `id` INT NOT NULL AUTO_INCREMENT,
-  `userId` INT NOT NULL,
+  `userId` BIGINT NOT NULL,
   `userName` VARCHAR(100) NOT NULL,
-  `tweetId` INT NOT NULL,
+  `tweetId` BIGINT NOT NULL,
   `status` TEXT NOT NULL,
   `favDate` DATETIME NOT NULL,
   `unfavDate` DATETIME NULL,
